refactor internal logs subscription
