+++ src/trace.rs
@@ -1,4 +1,4 @@
-use crate::event::Event;
+use crate::event::LogEvent;
 use metrics_tracing_context::MetricsLayer;
 use once_cell::sync::OnceCell;
 use std::sync::{Mutex, MutexGuard};
@@ -17,12 +17,12 @@ use tracing_subscriber::layer::SubscriberExt;
 /// BUFFER contains all of the internal log events generated by Vector
 /// before the topology has been initialized. It will be cleared (set to
 /// `None`) by the topology initialization routines.
-static BUFFER: OnceCell<Mutex<Option<Vec<Event>>>> = OnceCell::new();
+static BUFFER: OnceCell<Mutex<Option<Vec<LogEvent>>>> = OnceCell::new();
 
 /// SENDER holds the sender/receiver handle that will receive a copy of
 /// all the internal log events *after* the topology has been
 /// initialized.
-static SENDER: OnceCell<Sender<Event>> = OnceCell::new();
+static SENDER: OnceCell<Sender<LogEvent>> = OnceCell::new();
 
 pub use tracing_futures::Instrument;
 pub use tracing_tower::{InstrumentableService, InstrumentedService};
@@ -129,7 +129,7 @@ pub fn reset_early_buffer() {
     *early_buffer() = Some(Vec::new());
 }
 
-fn early_buffer() -> MutexGuard<'static, Option<Vec<Event>>> {
+fn early_buffer() -> MutexGuard<'static, Option<Vec<LogEvent>>> {
     BUFFER
         .get()
         .expect("Internal logs buffer not initialized")
@@ -146,8 +146,8 @@ pub fn current_span() -> Span {
 }
 
 pub struct TraceSubscription {
-    pub buffer: Vec<Event>,
-    pub receiver: Receiver<Event>,
+    pub buffer: Vec<LogEvent>,
+    pub receiver: Receiver<LogEvent>,
 }
 
 pub fn subscribe() -> TraceSubscription {
